F do not check file exists when using INET

+++ ss_manager/utils/transporter.py
@@ -31,7 +31,7 @@ class TransporterBase:
         sock = socket.socket(self.conn_type, self.protocol)
         if character == RECEIVER:
             # receiver
-            if os.path.exists(self.address):
+            if ":" not in self.address and os.path.exists(self.address):
                 os.remove(self.address)
             sock.bind(self.address)
         else:
